refactor: 通通精简掉，跑不通再说
