+++ requirements.txt
@@ -2,5 +2,3 @@ fastapi==0.104.1
 uvicorn==0.23.2
 pydantic==2.4.2
 pydantic-extra-types==2.1.0
-email-validator==2.1.0.post1
-python-multipart==0.0.6
