+++ src/auth.py
@@ -11,6 +11,45 @@ from const import CONFIG_FILENAME, ROOT_PATH
 from log import logger
 
 
+###
+
+from fastapi import Depends, HTTPException, Security, status
+from fastapi.security import APIKeyHeader
+from typing import Optional
+
+# In a real application, this would be stored securely
+API_KEYS = {"test_api_key": "admin"}
+
+api_key_header = APIKeyHeader(name="Authorization", auto_error=False)
+
+
+async def get_api_key(api_key_header: str = Security(api_key_header)) -> str:
+    if api_key_header is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Missing API Key",
+        )
+    
+    # Check if the header has the correct format (Bearer <token>)
+    if not api_key_header.startswith("Bearer "):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid API Key format. Use 'Bearer <api_key>'",
+        )
+    
+    # Extract the token
+    api_key = api_key_header.replace("Bearer ", "")
+    
+    if api_key not in API_KEYS:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid API Key",
+        )
+    
+    return api_key
+
+###
+
 def get_token() -> str:
     # 获取gist访问的时候需要一个github的token
