--- src/middleware/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the middleware directory a Python package
