--- src/models/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the models directory a Python package
