--- src/routes/__init__.py
@@ -1,2 +1,0 @@
-# 文件说明: 路由包初始化，导入字典路由 / File Description: Routes package initializer, imports dictionary router
-# This file makes the routes directory a Python package
